1.最新季的日期做排序 年->月番 (1月、4月、7月、10月) 2.依照系列做歸類，可以了解到說同一部出了多少季，所以之前的舊番可能會因為出新季度而被往前拉 3.日期選用最新集數的開播日期
