+++ src/TVshows.func.ts
@@ -12,7 +12,7 @@ const query = {
 };
 
 export default async (keywords: string[]): Promise<BaseSeries[]> => {
-  const CACHE: BaseSeries[] = [];
+  const result: BaseSeries[] = [];
   let GLOBAL_COUNTER = 1;
   //轉變keywords語法
   let keySTR = '';
@@ -34,14 +34,29 @@ export default async (keywords: string[]): Promise<BaseSeries[]> => {
     }
     const resList = data['results'];
     for (const ii of resList) {
-      CACHE.push(ii);
+      result.push({
+        id: ii['id'],
+        title: ii['name'],
+        poster_path: ii['poster_path'],
+        adult: ii['adult'] !== undefined,
+        genre_ids: ii['genre_ids'],
+        first_air_date: ii['first_air_date'],
+        backdrop_path: ii['backdrop_path'],
+        original_language: ii['original_language'],
+        original_title: ii['original_name'],
+        Seasons: [],
+        popularity: ii['popularity'],
+        overview: ii['overview'],
+        vote_average: ii['vote_average'],
+        vote_count: ii['vote_count'],
+      });
       console.log(GLOBAL_COUNTER++ + '/' + data?.['total_results']);
     }
     cur_page++;
     MaxPage = data['total_pages'] > 1 ? data['total_pages'] : -1;
     if (process.env['MODE'] === 'DEBUG') break;
   }
-  return CACHE;
+  return result;
 };
 
 //
