+++ src/html/HEAD.ts
@@ -1,0 +1,10 @@
+import STYLE from './STYLE';
+
+export default (): string => {
+  return `
+    <head>
+    <title>Movie-TV-Generater</title>
+      ${STYLE}
+    </head>
+    `;
+};
